+++ backend/src/enums/status-notificacao-enum.js
@@ -1,0 +1,4 @@
+exports.values = {
+  Encerrada: 'ENCERRADA',
+  Excluida: 'EXCLUIDA',
+};
