+++ backend/src/resource/notificacao-resource.js
@@ -457,120 +457,3 @@ exports.excluirLoteLogicamenteNotificacao = async (req, res, next) => {
     return next(err);
   }
 };
-
-exports.consultarNotificacaoEvolucao = async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const notificacaoEvolucao = await models.Notificacao.findOne({
-      where: { id },
-      attributes: ['id', 'status'],
-      include: [{
-        model: models.Pessoa,
-        attributes: ['nome', 'tipoDocumento', 'numeroDocumento', 'telefoneResidencial', 'telefoneContato', 'telefoneCelular'],
-      },
-      { model: models.NotificacaoEvolucao },
-      ],
-    });
-    if (!notificacaoEvolucao) res.status(404).json({ error: 'Notificação não encontrada.' });
-    if (notificacaoEvolucao.status !== 'ABERTA') res.status(400).json({ error: 'Notificação não está mais aberta.' });
-
-    return res.json({ data: notificacaoEvolucao });
-  } catch (err) {
-    return next(err);
-  }
-};
-
-const validarNotificacaoFinalizada = async (evolucao) => {
-  const notificacaoFinalizada = await models.Notificacao.findOne({
-    attributes: ['status'],
-    where: {
-      id: evolucao.notificacaoId,
-      [Op.or]: [{ status: 'ENCERRADA' }, { status: 'EXCLUIDA' }],
-    },
-  });
-
-  if (notificacaoFinalizada) {
-    throw new RegraNegocioErro(`Não é possível adicionar nova evolução pois a notificação está ${
-      notificacaoFinalizada.status}.`);
-  }
-};
-
-const validarPossuiConfirmacao = async (evolucao) => {
-  const tpEvolucaoPrecisaTerConfirmacao = (evolucao.tpEvolucao === 'CURA'
-    || evolucao.tpEvolucao === 'OBITO');
-
-  const tpEvolucaoProibidaSeJaConfirmada = (evolucao.tpEvolucao === 'SUSPEITO'
-    || evolucao.tpEvolucao === 'DESCARTADO'
-    || evolucao.tpEvolucao === 'CONFIRMADO'
-    || evolucao.tpEvolucao === 'ENCERRADO');
-
-  if (!(tpEvolucaoPrecisaTerConfirmacao || tpEvolucaoProibidaSeJaConfirmada)) {
-    return;
-  }
-
-  const evolucaoConfirmado = await models.NotificacaoEvolucao.findOne({
-    attributes: ['tpEvolucao'],
-    where: {
-      notificacaoId: evolucao.notificacaoId,
-      tpEvolucao: 'CONFIRMADO',
-    },
-  });
-
-  if (evolucaoConfirmado && tpEvolucaoProibidaSeJaConfirmada) {
-    const msgErro = `Não é possivel atualizar para ${evolucao.tpEvolucao} pois já existe atualização de confirmação.`;
-    throw new RegraNegocioErro(msgErro);
-  }
-
-  if (!evolucaoConfirmado && tpEvolucaoPrecisaTerConfirmacao) {
-    throw new RegraNegocioErro(`Não é possivel atualizar para ${evolucao.tpEvolucao}
-      pois não existe atualização de confirmação.`);
-  }
-};
-
-const encerrarNotificacao = async (evolucao, t) => {
-  const deveEncerrar = (evolucao.tpEvolucao === 'CURA'
-    || evolucao.tpEvolucao === 'DESCARTADO'
-    || evolucao.tpEvolucao === 'ENCERRADO'
-    || evolucao.tpEvolucao === 'OBITO');
-
-  if (!deveEncerrar) {
-    return;
-  }
-
-  await models.Notificacao.update(
-    { status: 'ENCERRADA' },
-    {
-      where: { id: evolucao.notificacaoId },
-      transaction: t,
-    },
-  );
-};
-
-const validarDataEvolucaoSuperiorDataNotificacao = async ({ notificacaoId, dtEvolucao }) => {
-  const { NotificacaoCovid19 } = await consultarNotificacaoPorId(notificacaoId);
-  const dataHoraNotificacao = new Date(NotificacaoCovid19.dataHoraNotificacao);
-  const dataEvolucao = new Date(dtEvolucao);
-  if (!NotificacaoCovid19) throw new Error(`Não foi possível encontrar a notificação ${notificacaoId}`);
-  if (dataEvolucao < dataHoraNotificacao) throw new RegraNegocioErro('A data da evolução não pode ser menor que a data da notificação.');
-};
-
-exports.salvarEvolucao = async (req, res, next) => {
-  try {
-    const result = await models.sequelize.transaction(async (t) => {
-      const evolucaoReq = req.body;
-
-      await validarDataEvolucaoSuperiorDataNotificacao(evolucaoReq);
-      await validarNotificacaoFinalizada(evolucaoReq);
-      await validarPossuiConfirmacao(evolucaoReq);
-
-      const evolucao = await models.NotificacaoEvolucao.create(evolucaoReq, { transaction: t });
-
-      await encerrarNotificacao(evolucaoReq, t);
-      return evolucao;
-    });
-
-    return res.json({ data: result });
-  } catch (err) {
-    return next(err);
-  }
-};
