+++ backend/src/routes/notification-route.js
@@ -2,6 +2,7 @@ const express = require('express');
 
 const router = express.Router();
 const NotificacaoResource = require('../resource/notificacao-resource');
+const NotificacaoEvolucaoResource = require('../resource/notificacao-evolucao-resource');
 const { validate, schemas } = require('../validations');
 
 const prefixoRoute = '/notificacoes';
@@ -21,7 +22,8 @@ router.get(`${prefixoRoute}/:id`, NotificacaoResource.consultarPorId);
 router.delete(`${prefixoRoute}/:id`, NotificacaoResource.excluirLogicamenteNotificacao);
 router.delete(`${prefixoRoute}`, NotificacaoResource.excluirLoteLogicamenteNotificacao);
 
-router.get(`${prefixoRoute}/:id/evolucoes`, NotificacaoResource.consultarNotificacaoEvolucao);
-router.post(`${prefixoRoute}:id/evolucoes`, NotificacaoResource.salvarEvolucao);
+router.get(`${prefixoRoute}/:id/evolucoes`, NotificacaoEvolucaoResource.consultar);
+router.post(`${prefixoRoute}/:id/evolucoes`, NotificacaoEvolucaoResource.cadastrar);
+router.delete(`${prefixoRoute}/:notificacaoId/evolucoes/:id`, NotificacaoEvolucaoResource.deletar);
 
 module.exports = router;
